+++ packages/snap-url-manager/src/Translators/Url/UrlTranslator.ts
@@ -148,27 +148,37 @@ export class UrlTranslator implements Translator {
 			.split('&')
 			.filter((v) => v)
 			.map((kvPair) => {
-				const [key, value] = kvPair.split('=').map((v) => decodeURIComponent(v.replace(/\+/g, ' ')));
-				return { key: key.split('.'), value, type: ParamLocationType.query };
+				try {
+					const [key, value] = kvPair.split('=').map((v) => decodeURIComponent(v.replace(/\+/g, ' ')));
+					return { key: key.split('.'), value, type: ParamLocationType.query };
+				} catch (err) {
+					console.warn('Snap UrlTranslator: URI malformed - ignoring parameter', kvPair);
+					return { key: ['ss__delete'], value: 'ss__delete', type: ParamLocationType.query };
+				}
 			})
 			.filter((param) => {
-				// remove core fields that do not contain a value
+				// remove core fields that do not contain a value or had malformed data
 				const isCoreField = this.reverseMapping[param.key[0]];
-				return !isCoreField || (isCoreField && param.value);
+				return param.value !== 'ss__delete' ? !isCoreField || (isCoreField && param.value) : '';
 			});
 	}
 
 	protected parseHashString(hashString: string): Array<UrlParameter> {
 		const params: Array<UrlParameter> = [];
-		const justHashString = hashString.split('#').join('/') || '';
+		const justHashString = hashString.split('#').pop() || '';
 		justHashString
 			.split('/')
 			.filter((v) => v)
 			.map((hashEntries) => {
-				return hashEntries.split(':').map((v) => decodeHashComponent(v));
+				try {
+					return hashEntries.split(':').map((v) => decodeHashComponent(v));
+				} catch (err) {
+					console.warn('Snap UrlTranslator: URI malformed - ignoring parameter', hashEntries);
+					return [];
+				}
 			})
 			.filter((param) => {
-				// remove core fields that do not contain a value
+				// remove core fields that do not contain a value or had malformed data
 				const [key, value] = param;
 				const isCoreField = this.reverseMapping[key];
 				return !isCoreField || (isCoreField && value);